Refactor code
- Clean unused user modules
- Refactor Github code
- Fix warnings
