--- lib/orange_cms/accounts/user.ex
@@ -1,18 +1,0 @@
-defmodule OrangeCms.Accounts.User do
-  use Ecto.Schema
-  import Ecto.Changeset
-
-  schema "users" do
-    field :age, :integer
-    field :name, :string
-
-    timestamps()
-  end
-
-  @doc false
-  def changeset(user, attrs) do
-    user
-    |> cast(attrs, [:name, :age])
-    |> validate_required([:name, :age])
-  end
-end
