+++ lib/orange_cms_web/live/content_entry_live/edit.html.heex
@@ -29,7 +29,7 @@
                 <% frontmatter = Phoenix.HTML.Form.input_value(f, :frontmatter) %>
                 <div>
                     <div :for={field_def <- @content_type.field_defs} class="mb-4">
-                         <.schema_field field={field_def}, value={frontmatter[field_def.key]} data={%{options: [prefix: "frontmatter"], project: @current_project}}/>
+                         <.schema_field field={field_def} value={frontmatter[field_def.key]} data={%{options: [prefix: "frontmatter"], project: @current_project}}/>
                     </div>
                 </div>
             </div>
@@ -38,7 +38,7 @@
                 <div id="md-editor-container" phx-update="ignore">
                     <label class="sr-only" for="message">Message</label>
 
-                    <.input type="textarea" field={f[:raw_body]} class="w-full rounded-lg border-gray-200 p-3 text-sm" rows="30" id="md-editor" phx-hook="MdEditor" data-upload-path={~p"/app/api/upload_image/#{@current_project.id}"} />
+                    <.input type="textarea" field={f[:raw_body]} class="w-full rounded-lg border-gray-200 p-3 text-sm" rows="30" id="md-editor" phx-hook="MdEditor" data-upload-path={~p"/app/api/upload_image/#{@current_project.id}/#{@current_type.id}"} />
                 </div>
 
                 <div class="mt-4">
