+++ lib/orange_cms_web/live/project_live/index.html.heex
@@ -2,7 +2,7 @@
     Listing Projects
     <:actions>
         <.link patch={~p"/app/p/new"}>
-            <.button>New Project</.button>
+            <.button class="btn btn-md btn-secondary"> <Heroicons.plus class="w-5 h-5" /> New project</.button>
         </.link>
     </:actions>
 </.header>
