--- lib/orange_cms_web/live/user_live/form_component.ex
@@ -1,91 +1,0 @@
-defmodule OrangeCmsWeb.UserLive.FormComponent do
-  use OrangeCmsWeb, :live_component
-
-  alias OrangeCms.Accounts
-
-  @impl true
-  def render(assigns) do
-    ~H"""
-    <div>
-      <.header>
-        <%= @title %>
-        <:subtitle>Use this form to manage user records in your database.</:subtitle>
-      </.header>
-
-      <.simple_form
-        for={@form}
-        id="user-form"
-        phx-target={@myself}
-        phx-change="validate"
-        phx-submit="save"
-      >
-        <.input field={@form[:name]} type="text" label="Name" />
-        <.input field={@form[:age]} type="number" label="Age" />
-        <:actions>
-          <.button phx-disable-with="Saving...">Save User</.button>
-        </:actions>
-      </.simple_form>
-    </div>
-    """
-  end
-
-  @impl true
-  def update(%{user: user} = assigns, socket) do
-    changeset = Accounts.change_user(user)
-
-    {:ok,
-     socket
-     |> assign(assigns)
-     |> assign_form(changeset)}
-  end
-
-  @impl true
-  def handle_event("validate", %{"user" => user_params}, socket) do
-    changeset =
-      socket.assigns.user
-      |> Accounts.change_user(user_params)
-      |> Map.put(:action, :validate)
-
-    {:noreply, assign_form(socket, changeset)}
-  end
-
-  def handle_event("save", %{"user" => user_params}, socket) do
-    save_user(socket, socket.assigns.action, user_params)
-  end
-
-  defp save_user(socket, :edit, user_params) do
-    case Accounts.update_user(socket.assigns.user, user_params) do
-      {:ok, user} ->
-        notify_parent({:saved, user})
-
-        {:noreply,
-         socket
-         |> put_flash(:info, "User updated successfully")
-         |> push_patch(to: socket.assigns.patch)}
-
-      {:error, %Ecto.Changeset{} = changeset} ->
-        {:noreply, assign_form(socket, changeset)}
-    end
-  end
-
-  defp save_user(socket, :new, user_params) do
-    case Accounts.create_user(user_params) do
-      {:ok, user} ->
-        notify_parent({:saved, user})
-
-        {:noreply,
-         socket
-         |> put_flash(:info, "User created successfully")
-         |> push_patch(to: socket.assigns.patch)}
-
-      {:error, %Ecto.Changeset{} = changeset} ->
-        {:noreply, assign_form(socket, changeset)}
-    end
-  end
-
-  defp assign_form(socket, %Ecto.Changeset{} = changeset) do
-    assign(socket, :form, to_form(changeset))
-  end
-
-  defp notify_parent(msg), do: send(self(), {__MODULE__, msg})
-end
