+++ lib/orange_cms_web/plug/load_project.ex
@@ -2,7 +2,8 @@ defmodule OrangeCmsWeb.LoadProject do
   @moduledoc """
   Ensures common `assigns` are applied to all LiveViews attaching this hook.
   """
-  use OrangeCmsWeb, :live_view
+  use OrangeCmsWeb, :verified_routes
+  import Phoenix.LiveView, only: [push_navigate: 2, put_flash: 3]
   alias OrangeCms.Projects.Project
 
   def on_mount(_, params, _session, socket) do
