+++ lib/orange_cms_web/plug/menu_assign.ex
@@ -2,8 +2,6 @@ defmodule OrangeCmsWeb.MenuAssign do
   @moduledoc """
   Ensures common `assigns` are applied to all LiveViews attaching this hook.
   """
-  import Phoenix.LiveView
-  alias OrangeCms.Content
   alias OrangeCms.Content.ContentType
 
   def on_mount(_, params, _session, socket) do
@@ -22,7 +20,7 @@ defmodule OrangeCmsWeb.MenuAssign do
            current_type: current_type
          )}
 
-      {:error, err} ->
+      {:error, _err} ->
         {:halt, socket}
     end
   end
